requirements.txt
discord.py == 2.3.1
requests == 2.26.0
simplejson == 3.17.3